backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
marshmallow==3.20.1
python-dotenv==1.0.0
boto3==1.34.0
requests==2.31.0
pyyaml==6.0.1
schedule==1.2.0
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4 